+++ packages/sequelize/tests/models/IOrderingItem.ts
@@ -1,0 +1,4 @@
+export interface IOrderingItem {
+  id: number;
+  order?: number;
+}