+++ packages/sequelize/tests/models/ProviderCity.ts
@@ -1,4 +1,0 @@
-export interface IProviderCity {
-  cityId: number;
-  providerId: number;
-}