+++ packages/sequelize/tests/repositories/CitiesRepository.ts
@@ -1,4 +1,4 @@
 import { BaseRepository } from '../../src';
-import { City, ICityEntity, ICity } from '../models';
+import { City, ICityEntity, ICityWithAssociations, IOrderingItem } from '../models';
 
-export const CitiesRepository = new BaseRepository<ICityEntity, ICity>(City);
+export const CitiesRepository = new BaseRepository<ICityEntity, ICityWithAssociations, IOrderingItem>(City);