+++ packages/sequelize/tests/repositories/ProvidersRepository.ts
@@ -1,4 +1,4 @@
 import { BaseRepository } from '../../src';
-import { Provider, IProviderEntity, IProvider } from '../models';
+import { Provider, IProviderEntity, IProvider, IOrderingItem } from '../models';
 
-export const ProvidersRepository = new BaseRepository<IProviderEntity, IProvider>(Provider);
+export const ProvidersRepository = new BaseRepository<IProviderEntity, IProvider, IOrderingItem>(Provider);